Applies template pattern to reduce duplicate code in smartmetering responseMessageProcessors.
