+++ osgp-adapter-domain-smartmetering/src/main/java/com/alliander/osgp/adapter/domain/smartmetering/infra/jms/core/messageprocessors/ActualMeterReadsResponseMessageProcessor.java
@@ -7,11 +7,6 @@
  */
 package com.alliander.osgp.adapter.domain.smartmetering.infra.jms.core.messageprocessors;
 
-import javax.jms.JMSException;
-import javax.jms.ObjectMessage;
-
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
@@ -21,14 +16,10 @@ import com.alliander.osgp.domain.core.valueobjects.DeviceFunction;
 import com.alliander.osgp.dto.valueobjects.smartmetering.ActualMeterReads;
 import com.alliander.osgp.dto.valueobjects.smartmetering.MeterReadsGas;
 import com.alliander.osgp.shared.exceptionhandling.OsgpException;
-import com.alliander.osgp.shared.infra.jms.Constants;
 import com.alliander.osgp.shared.infra.jms.ResponseMessage;
-import com.alliander.osgp.shared.infra.jms.ResponseMessageResultType;
 
 @Component("domainSmartMeteringActualMeterReadsResponseMessageProcessor")
 public class ActualMeterReadsResponseMessageProcessor extends OsgpCoreResponseMessageProcessor {
-
-    private static final Logger LOGGER = LoggerFactory.getLogger(ActualMeterReadsResponseMessageProcessor.class);
 
     @Autowired
     private MonitoringService monitoringService;
@@ -38,55 +29,19 @@ public class ActualMeterReadsResponseMessageProcessor extends OsgpCoreResponseMe
     }
 
     @Override
-    public void processMessage(final ObjectMessage message) throws JMSException {
-        LOGGER.debug("Processing smart metering default response message");
+    protected void handleMessage(final String deviceIdentification, final String organisationIdentification,
+            final String correlationUid, final String messageType, final ResponseMessage responseMessage,
+            final OsgpException osgpException) {
 
-        String correlationUid = null;
-        String messageType = null;
-        String organisationIdentification = null;
-        String deviceIdentification = null;
+        if (responseMessage.getDataObject() instanceof ActualMeterReads) {
+            final ActualMeterReads actualMeterReadsDto = (ActualMeterReads) responseMessage.getDataObject();
 
-        ResponseMessage responseMessage = null;
-        ResponseMessageResultType responseMessageResultType = null;
-        OsgpException osgpException = null;
-
-        try {
-            correlationUid = message.getJMSCorrelationID();
-            messageType = message.getJMSType();
-            organisationIdentification = message.getStringProperty(Constants.ORGANISATION_IDENTIFICATION);
-            deviceIdentification = message.getStringProperty(Constants.DEVICE_IDENTIFICATION);
-
-            responseMessage = (ResponseMessage) message.getObject();
-            responseMessageResultType = responseMessage.getResult();
-            osgpException = responseMessage.getOsgpException();
-        } catch (final JMSException e) {
-            LOGGER.error("UNRECOVERABLE ERROR, unable to read ObjectMessage instance, giving up.", e);
-            LOGGER.debug("correlationUid: {}", correlationUid);
-            LOGGER.debug("messageType: {}", messageType);
-            LOGGER.debug("organisationIdentification: {}", organisationIdentification);
-            LOGGER.debug("deviceIdentification: {}", deviceIdentification);
-            LOGGER.debug("responseMessageResultType: {}", responseMessageResultType);
-            LOGGER.debug("deviceIdentification: {}", deviceIdentification);
-            LOGGER.debug("osgpException: {}", osgpException);
-            return;
-        }
-
-        try {
-            LOGGER.info("Calling application service function to handle response: {}", messageType);
-
-            if (responseMessage.getDataObject() instanceof ActualMeterReads) {
-                final ActualMeterReads actualMeterReadsDto = (ActualMeterReads) responseMessage.getDataObject();
-
-                this.monitoringService.handleActualMeterReadsResponse(deviceIdentification, organisationIdentification,
-                        correlationUid, messageType, responseMessageResultType, osgpException, actualMeterReadsDto);
-            } else {
-                final MeterReadsGas meterReadsGas = (MeterReadsGas) responseMessage.getDataObject();
-                this.monitoringService.handleActualMeterReadsResponse(deviceIdentification, organisationIdentification,
-                        correlationUid, messageType, responseMessageResultType, osgpException, meterReadsGas);
-            }
-
-        } catch (final Exception e) {
-            this.handleError(e, correlationUid, organisationIdentification, deviceIdentification, messageType);
+            this.monitoringService.handleActualMeterReadsResponse(deviceIdentification, organisationIdentification,
+                    correlationUid, messageType, responseMessage.getResult(), osgpException, actualMeterReadsDto);
+        } else {
+            final MeterReadsGas meterReadsGas = (MeterReadsGas) responseMessage.getDataObject();
+            this.monitoringService.handleActualMeterReadsResponse(deviceIdentification, organisationIdentification,
+                    correlationUid, messageType, responseMessage.getResult(), osgpException, meterReadsGas);
         }
     }
